requirements.txt
# Core LLM and AI dependencies
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.5

# Vector database and RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Language processing
spacy>=3.7.0
googletrans>=4.0.0
polyglot>=16.7.4

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Database and storage
sqlalchemy>=2.0.0

# API and web framework
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and monitoring
loguru>=0.7.0

# Scientific validation
scikit-learn>=1.3.0
nltk>=3.8.0